Trabalho 3 corrigir dtype do vetor de mate

+++ grafo.py
@@ -39,7 +39,7 @@ class Grafo:
                     self.graus[b-1] += 1
                 else:
                     print(f"Aresta ({a}, {b}) já existe!")
-                self.add_neighbour(int(a), int(b))
+                self.add_neighbour(a, b)
 
     # Vizualizing
     def mostrar_grafo(self) -> None:
@@ -496,7 +496,7 @@ class GrafoBipartido(GrafoNaoDirigido):
     
     def hopcroft_karp(self):
         D = np.ones(self.qtdVertices())*np.inf
-        mate = np.ones(self.qtdVertices())*-1
+        mate = np.ones(self.qtdVertices(), dtype=int)*-1
         m = 0
         while self.hopcroft_karp_bfs(mate, D):
             for x in self.X:
@@ -514,20 +514,20 @@ class GrafoBipartido(GrafoNaoDirigido):
                 D[x] = np.inf
         D[-1] = np.inf
         while not Q.empty():
-            x = int(Q.get())
-            if D[int(x)] < D[-1]:
+            x = Q.get()
+            if D[x] < D[-1]:
                 for y in self.vizinhos(x):
-                    if D[int(mate[y])] == np.inf:
-                        D[int(mate[y])] = D[x] + 1
+                    if D[mate[y]] == np.inf:
+                        D[mate[y]] = D[x] + 1
                         Q.put(mate[y])
         return D[-1] != np.inf
     
     def hopcroft_karp_dfs(self, x, mate, D):
         if x != -1:
             for y in range(self.qtdVertices()):
-                if self[x, y] != np.inf and x != y and D[int(mate[y])] == D[int(x)] + 1 and self.hopcroft_karp_dfs(mate[y], mate, D):
+                if self[x, y] != np.inf and x != y and D[mate[y]] == D[x] + 1 and self.hopcroft_karp_dfs(mate[y], mate, D):
                     mate[y] = x
-                    mate[int(x)] = y
+                    mate[x] = y
                     return True
             D[x] = np.inf
             return False
